Create blueprint applications auth, reports, main.
