+++ app/__init__.py
@@ -5,21 +5,51 @@ from flask_migrate import Migrate
 from flask_login import LoginManager
 from flask_bootstrap import Bootstrap
 
+import logging
+from logging.handlers import RotatingFileHandler
+import os
 
-app = Flask(__name__)
-app.config.from_object(Config)
-db = SQLAlchemy(app)
-migrate = Migrate(app, db)
-login = LoginManager(app)
+
+db = SQLAlchemy()
+migrate = Migrate()
+login = LoginManager()
 login.login_view = 'index'
-bootstrap = Bootstrap(app)
+bootstrap = Bootstrap()
 
 
-if __name__ == '__main__':
-    app.run(debug=True)
+def create_app(config_class=Config):
+
+    app = Flask(__name__)
+    app.config.from_object(config_class)
+    db.init_app(app)
+    migrate.init_app(app, db)
+    login.init_app(app)
+    bootstrap.init_app(app)
+
+    from app.auth import bp as auth_bp
+    app.register_blueprint(auth_bp)
+    from app.reports import bp as reports_bp
+    app.register_blueprint(reports_bp)
+    from app.main import bp as main_bp
+    app.register_blueprint(main_bp)
+
+    if not app.debug:
+
+        if not os.path.exists('logs'):
+            os.mkdir('logs')
+        file_handler = RotatingFileHandler('logs/affiliate.log', maxBytes=10240, backupCount=10)
+        file_handler.setFormatter(
+            logging.Formatter('%(asctime)s %(levelname)s: %(message)s [in %(pathname)s:%(lineno)d]'))
+        file_handler.setLevel(logging.INFO)
+        app.logger.addHandler(file_handler)
+
+        app.logger.setLevel(logging.INFO)
+        app.logger.info('Affiliate startup')
+
+    return app
 
 
-from app import models, routes
+from app import models
 
 
 @login.user_loader
