+++ app/reports/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('reports', __name__)
+
+from app.reports import routes
