--- app/routes.py
@@ -1,223 +1,0 @@
-from flask import render_template, redirect, url_for, flash, session, request
-from flask_login import current_user, login_user, login_required, logout_user
-from flask.views import MethodView
-from app.forms import LoginForm, RegistrationForm, LinkForm, PersonalDataEditForm, ChangePasswordForm
-from app import app, db
-from app.models import User, Link, Action
-from datetime import datetime
-from sqlalchemy import func, distinct
-
-
-@app.route('/')
-def index():
-    return render_template('index.html')
-
-
-@app.route('/login', methods=['GET', 'POST'])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = LoginForm()
-
-    if form.validate_on_submit():
-        user = User.query.filter_by(email=form.email.data).first()
-
-        if not user or not user.check_password(form.password.data):
-            flash('User with provided credentials does not exist.')
-            return redirect(url_for('login'))
-
-        login_user(user, remember=form.remember_me.data)
-        return redirect(url_for('dashboard'))
-
-    return render_template('login.html', form=form)
-
-
-@app.route('/register', methods=['GET', 'POST'])
-def register():
-    if current_user.is_authenticated:
-        return redirect(url_for('index'))
-    form = RegistrationForm()
-
-    if form.validate_on_submit():
-        check_user = User.query.filter_by(email=form.email.data).first()
-        if not check_user:
-            user = User(email=form.email.data,
-                        first_name=form.first_name.data,
-                        last_name=form.last_name.data,
-                        messenger_type=form.messenger_type.data,
-                        messenger=form.messenger.data)
-
-            user.set_password(form.password.data)
-            db.session.add(user)
-            db.session.commit()
-            flash('Registration completed successfully')
-            return redirect(url_for('index'))
-        else:
-            flash('Sorry, email already exist.')
-    return render_template('register.html', form=form)
-
-
-@app.route('/logout')
-def logout():
-    logout_user()
-    return redirect(url_for('index'))
-
-
-@app.route('/change_password', methods=['get', 'post'])
-@login_required
-def change_password():
-    user = User.query.get(current_user.id)
-    form = ChangePasswordForm()
-    if form.validate_on_submit():
-        user.set_password(form.password.data)
-        db.session.add(user)
-        db.session.commit()
-        return redirect('/profile')
-    return render_template('change_password.html', form=form)
-
-
-@app.route('/dashboard')
-@login_required
-def dashboard():
-    user = current_user
-    data = (
-        Action.query.
-        join(Link).
-        filter_by(user_id=user.id).
-        with_entities(func.count(Action.ip_address).label('ip'),
-                      func.count(distinct(Action.ip_address)).label('unique'),
-                      func.count(Action.purchase_amount).label('purchases'),
-                      func.sum(Action.purchase_amount).label('amount')).first())
-    return render_template('dashboard.html', user=user, data=data)
-
-
-class ProfileView(MethodView):
-
-    @login_required
-    def get(self):
-        form = PersonalDataEditForm()
-        form.email.data = current_user.email
-        form.first_name.data = current_user.first_name
-        form.last_name.data = current_user.last_name
-        form.messenger_type.data = current_user.messenger_type
-        form.messenger.data = current_user.messenger
-        links = Link.query.filter_by(user_id=current_user.id).all()
-        return render_template('profile.html', form=form, links=links)
-
-    @login_required
-    def post(self):
-        form = PersonalDataEditForm()
-        if form.validate_on_submit():
-            current_user.email = form.email.data
-            current_user.first_name = form.first_name.data
-            current_user.last_name = form.last_name.data
-            current_user.messenger_type = form.messenger_type.data
-            current_user.messenger = form.messenger.data
-            db.session.commit()
-            flash('Your changes have been saved')
-            return redirect(url_for('profile'))
-
-
-@app.route('/links', methods=['get', 'post'])
-@login_required
-def links():
-    links_list = Link.query.filter_by(user_id=current_user.id).order_by(Link.timestamp.desc()).all()
-    form = LinkForm()
-    if form.validate_on_submit():
-        link = Link(name=form.name.data, site=form.site.data, user_id=current_user.id)
-        link.generate_hash()
-        db.session.add(link)
-        db.session.commit()
-        return redirect(url_for('links'))
-
-    return render_template('links.html', form=form, links_list=links_list)
-
-
-@app.route('/redirect_link/<hash>')
-def redirect_link(hash):
-    link = Link.query.filter_by(hash_str=hash).first()
-    if link:
-        ip = request.remote_addr
-        agent = request.headers.get('User-Agent')
-        new_action = Action(link_id=link.id, type_id=1, ip_address=ip, user_agent=agent, timestamp=datetime.utcnow())
-        db.session.add(new_action)
-        db.session.commit()
-
-    return redirect(link.site)
-
-
-@app.route('/reports/all_actions_report')
-@login_required
-def all_actions_report():
-    actions = (
-        Action.query
-        .join(Link)
-        .filter_by(user_id=current_user.id)
-        .order_by(Action.timestamp.desc()).all()
-    )
-    return render_template('all_actions_report.html', actions=actions)
-
-
-@app.route('/reports/all_links_report')
-@login_required
-def all_links_report():
-    actions = (
-        Action.query.
-        join(Link).
-        filter_by(user_id=current_user.id).
-        with_entities(Link,
-                      func.count(Action.ip_address).label('clicks'),
-                      func.count(distinct(Action.ip_address)).label('unique'),
-                      func.count(Action.purchase_amount).label('purchases'),
-                      func.sum(Action.purchase_amount).label('amount')).
-        group_by(Action.link_id).all()
-    )
-    return render_template('all_links_report.html', actions=actions)
-
-
-@app.route('/reports/current_link_report/<link_id>')
-@login_required
-def current_link_report(link_id):
-    link = Link.query.filter_by(id=link_id, user_id=current_user.id).first()
-    if not link:
-        return redirect(url_for('index'))
-    dates = (
-        Action.query.
-        join(Link).
-        filter_by(id=link_id).
-        with_entities(func.strftime("%Y-%m-%d", Action.timestamp).label('date'),
-                      func.count(Action.ip_address).label('actions'),
-                      func.count(distinct(Action.ip_address)).label('unique'),
-                      func.sum(Action.purchase_amount).label('purchases'),
-                      func.count(Action.purchase_amount).label('amount')).
-        group_by(func.strftime("%Y-%m-%d",  Action.timestamp)).all())
-    return render_template('current_link_report.html', dates=dates, link=link)
-
-
-@app.route('/reports/all_days_report')
-@login_required
-def all_days_report():
-    links = Link.query.filter_by(user_id=current_user.id).all()
-    links_id = [link.id for link in links]
-
-    sub = (
-        Action.query.
-        filter(Action.link_id.in_(links_id)).
-        with_entities(Action.ip_address.label('ip'),
-                      Action.link_id.label('name'),
-                      func.strftime('%Y-%m-%d', Action.timestamp).label('date')).
-        group_by(Action.ip_address, Action.link_id).subquery())
-
-    one = db.session.query(func.count(sub.c.ip).label('ip'), sub.c.date.label('date')).subquery()
-
-    dates = (
-        db.session.query(func.count(Action.ip_address).label('ip'),
-                         one.c.ip.label('unique'),
-                         one.c.date.label('date'),
-                         func.sum(Action.purchase_amount).label('purchases'),
-                         func.count(Action.purchase_amount).label('amount')).
-        group_by(func.strftime('%Y-%m-%d', Action.timestamp)).all())
-    return render_template('all_days_report.html', dates=dates)
-
-
-app.add_url_rule('/profile', view_func=ProfileView.as_view('profile'))
