+++ app/templates/auth/change_password.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'main/base.html' %}
 {% from 'bootstrap/form.html' import render_form %}
 
 {% block content %}
