+++ app/templates/main/base.html
@@ -10,28 +10,28 @@
 <body>
 
 <nav class="navbar navbar-expand-lg navbar-light bg-light mb-3">
-    <a class="navbar-brand" href="{{ url_for('index') }}">Affiliate program</a>
+    <a class="navbar-brand" href="{{ url_for('main.index') }}">Affiliate program</a>
     <div class="collapse navbar-collapse" id="navbarNav">
         <ul class="navbar-nav ml-auto">
             {% if current_user.is_authenticated %}
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('dashboard') }}">Dashboard</a>
+                    <a class="nav-link" href="{{ url_for('main.dashboard') }}">Dashboard</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('links') }}">Links</a>
+                    <a class="nav-link" href="{{ url_for('main.links') }}">Links</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('profile') }}">Profile</a>
+                    <a class="nav-link" href="{{ url_for('main.profile') }}">Profile</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('logout') }}">Logout</a>
+                    <a class="nav-link" href="{{ url_for('auth.logout') }}">Logout</a>
                 </li>
             {% else %}
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('login') }}">Sign In</a>
+                    <a class="nav-link" href="{{ url_for('auth.login') }}">Sign In</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="{{ url_for('register') }}">Sign Up</a>
+                    <a class="nav-link" href="{{ url_for('auth.register') }}">Sign Up</a>
                 </li>
             {% endif %}
         </ul>
