+++ app/templates/main/dashboard.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'main/base.html' %}
 
 {% block title %}{{ user.email }}{% endblock %}
