+++ app/templates/main/index.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'main/base.html' %}
 
 {% block content %}
     <p>Some additional information</p>
