+++ app/templates/main/profile.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %}
+{% extends 'main/base.html' %}
 {% import 'bootstrap/wtf.html' as wtf %}
 
 {% block content %}
@@ -13,7 +13,7 @@
                 {{ wtf.form_field(form.messenger_type) }}
                 {{ wtf.form_field(form.messenger) }}
                 <input type="submit" value="Save" class="btn btn-primary"><br>
-                <a href="{{ url_for('change_password') }}">Change password</a>
+                <a href="{{ url_for('auth.change_password') }}">Change password</a>
             </form>
             {% with messages = get_flashed_messages() %}
                 {% if messages %}
@@ -27,15 +27,15 @@
             <h3>Reports</h3>
             <h5>General:</h5>
             <ul class="list-group list-group-flush">
-                <li class="list-group-item"><a href="{{ url_for('all_actions_report') }}">All clicks</a></li>
-                <li class="list-group-item"><a href="{{ url_for('all_links_report') }}">All links</a></li>
-                <li class="list-group-item"><a href="{{ url_for('all_days_report') }}">All days</a></li>
+                <li class="list-group-item"><a href="{{ url_for('reports.all_actions_report') }}">All clicks</a></li>
+                <li class="list-group-item"><a href="{{ url_for('reports.all_links_report') }}">All links</a></li>
+                <li class="list-group-item"><a href="{{ url_for('reports.all_days_report') }}">All days</a></li>
             </ul>
             <p></p>
             <h5>Current link:</h5>
             <ul class="list-group list-group-flush">
                 {% for link in links %}
-                    <li class="list-group-item"><a href="{{ url_for('current_link_report', link_id=link.id) }}">{{ link.name }}</a></li>
+                    <li class="list-group-item"><a href="{{ url_for('reports.current_link_report', link_id=link.id) }}">{{ link.name }}</a></li>
                 {% endfor %}
             </ul>
         </div>
