+++ logs/affiliate.log.1
@@ -1,0 +1,81 @@
+2019-10-12 16:16:59,974 ERROR: Exception on /register [GET] [in /Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py:1891]
+Traceback (most recent call last):
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 2446, in wsgi_app
+    response = self.full_dispatch_request()
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1951, in full_dispatch_request
+    rv = self.handle_user_exception(e)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1820, in handle_user_exception
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1949, in full_dispatch_request
+    rv = self.dispatch_request()
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1935, in dispatch_request
+    return self.view_functions[rule.endpoint](**req.view_args)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/auth/routes.py", line 50, in register
+    return render_template('auth/register.html', form=form)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/templating.py", line 140, in render_template
+    ctx.app,
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/templating.py", line 120, in _render
+    rv = template.render(context)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/asyncsupport.py", line 76, in render
+    return original_render(self, *args, **kwargs)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/environment.py", line 1008, in render
+    return self.environment.handle_exception(exc_info, True)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/environment.py", line 780, in handle_exception
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/_compat.py", line 37, in reraise
+    raise value.with_traceback(tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/templates/auth/register.html", line 2, in top-level template code
+    {% from 'bootstrap/form.html' import render_form %}
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/templates/main/base.html", line 42, in top-level template code
+    {% block content %}
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/templates/auth/register.html", line 14, in block "content"
+    {{ render_form(form, action='', method='post', button_map={'submit': 'primary'}) }}
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/runtime.py", line 579, in _invoke
+    rv = self._func(*arguments)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask_bootstrap/templates/bootstrap/form.html", line 240, in template
+    {{ render_field(field,
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/runtime.py", line 579, in _invoke
+    rv = self._func(*arguments)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask_bootstrap/templates/bootstrap/form.html", line 163, in template
+    {{ field(class="form-control", **kwargs)|safe }}
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/wtforms/fields/core.py", line 155, in __call__
+    return self.meta.render_field(self, kwargs)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/wtforms/meta.py", line 56, in render_field
+    return field.widget(field, **render_kw)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/wtforms/widgets/core.py", line 323, in __call__
+    for val, label, selected in field.iter_choices():
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/wtforms/fields/core.py", line 454, in iter_choices
+    for value, label in self.choices:
+TypeError: 'NoneType' object is not iterable
+2019-10-12 16:17:12,425 ERROR: Exception on /links [POST] [in /Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py:1891]
+Traceback (most recent call last):
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 2446, in wsgi_app
+    response = self.full_dispatch_request()
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1951, in full_dispatch_request
+    rv = self.handle_user_exception(e)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1820, in handle_user_exception
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1949, in full_dispatch_request
+    rv = self.dispatch_request()
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1935, in dispatch_request
+    return self.view_functions[rule.endpoint](**req.view_args)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask_login/utils.py", line 261, in decorated_view
+    return func(*args, **kwargs)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/main/routes.py", line 69, in links
+    return redirect(url_for('auth.links'))
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/helpers.py", line 370, in url_for
+    return appctx.app.handle_url_build_error(error, endpoint, values)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 2215, in handle_url_build_error
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/helpers.py", line 358, in url_for
+    endpoint, values, method=method, force_external=external
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/werkzeug/routing.py", line 2020, in build
+    raise BuildError(endpoint, values, method, self)
+werkzeug.routing.BuildError: Could not build url for endpoint 'auth.links'. Did you mean 'auth.login' instead?
+2019-10-12 16:17:51,327 INFO: Affiliate startup [in /Users/zverev/PycharmProjects/learn_python/affiliate/app/__init__.py:47]
