+++ logs/affiliate.log.3
@@ -1,0 +1,74 @@
+2019-10-12 16:15:08,586 ERROR: Exception on /login [GET] [in /Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py:1891]
+Traceback (most recent call last):
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 2446, in wsgi_app
+    response = self.full_dispatch_request()
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1951, in full_dispatch_request
+    rv = self.handle_user_exception(e)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1820, in handle_user_exception
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1949, in full_dispatch_request
+    rv = self.dispatch_request()
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1935, in dispatch_request
+    return self.view_functions[rule.endpoint](**req.view_args)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/auth/routes.py", line 25, in login
+    return render_template('auth/login.html', form=form)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/templating.py", line 140, in render_template
+    ctx.app,
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/templating.py", line 120, in _render
+    rv = template.render(context)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/asyncsupport.py", line 76, in render
+    return original_render(self, *args, **kwargs)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/environment.py", line 1008, in render
+    return self.environment.handle_exception(exc_info, True)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/environment.py", line 780, in handle_exception
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/jinja2/_compat.py", line 37, in reraise
+    raise value.with_traceback(tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/templates/auth/login.html", line 2, in top-level template code
+    {% from 'bootstrap/form.html' import render_form %}
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/templates/main/base.html", line 42, in top-level template code
+    {% block content %}
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/templates/auth/login.html", line 21, in block "content"
+    <p>Do not have account?<a href="{{ url_for('register') }}"> Signup</a></p>
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/helpers.py", line 370, in url_for
+    return appctx.app.handle_url_build_error(error, endpoint, values)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 2215, in handle_url_build_error
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/helpers.py", line 358, in url_for
+    endpoint, values, method=method, force_external=external
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/werkzeug/routing.py", line 2020, in build
+    raise BuildError(endpoint, values, method, self)
+werkzeug.routing.BuildError: Could not build url for endpoint 'register'. Did you mean 'auth.register' instead?
+2019-10-12 16:15:22,165 INFO: Affiliate startup [in /Users/zverev/PycharmProjects/learn_python/affiliate/app/__init__.py:47]
+2019-10-12 16:15:29,510 ERROR: Exception on /login [POST] [in /Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py:1891]
+Traceback (most recent call last):
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 2446, in wsgi_app
+    response = self.full_dispatch_request()
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1951, in full_dispatch_request
+    rv = self.handle_user_exception(e)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1820, in handle_user_exception
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1949, in full_dispatch_request
+    rv = self.dispatch_request()
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 1935, in dispatch_request
+    return self.view_functions[rule.endpoint](**req.view_args)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/app/auth/routes.py", line 23, in login
+    return redirect(url_for('dashboard'))
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/helpers.py", line 370, in url_for
+    return appctx.app.handle_url_build_error(error, endpoint, values)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/app.py", line 2215, in handle_url_build_error
+    reraise(exc_type, exc_value, tb)
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/_compat.py", line 39, in reraise
+    raise value
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/flask/helpers.py", line 358, in url_for
+    endpoint, values, method=method, force_external=external
+  File "/Users/zverev/PycharmProjects/learn_python/affiliate/venv/lib/python3.7/site-packages/werkzeug/routing.py", line 2020, in build
+    raise BuildError(endpoint, values, method, self)
+werkzeug.routing.BuildError: Could not build url for endpoint 'dashboard'. Did you mean 'main.dashboard' instead?
+2019-10-12 16:16:24,923 INFO: Affiliate startup [in /Users/zverev/PycharmProjects/learn_python/affiliate/app/__init__.py:47]
